Add support for time-offsets that are floats (OD does that with some books, I guess, though it appears to be pretty infrequent)

+++ Book.py
@@ -72,7 +72,7 @@ class Book:
 				ch['od-path'] = None
 				ch['filename'] = None
 			if len(info) > 1:
-				ch['time-offset'] = int(info[1])
+				ch['time-offset'] = float(info[1])
 			else:
 				ch['time-offset'] = None
